+++ lib/presentation/blocs/authentication/authentication_state.dart
@@ -2,13 +2,13 @@ part of 'authentication_bloc.dart';
 
 @freezed
 sealed class AuthenticationState with _$AuthenticationState {
-  const factory AuthenticationState.initial() = _Initial;
+  const factory AuthenticationState.initial() = Initial;
 
-  const factory AuthenticationState.loading() = _Loading;
+  const factory AuthenticationState.loading() = Loading;
 
-  const factory AuthenticationState.emailNotVerified() = _EmailNotVerified;
+  const factory AuthenticationState.emailNotVerified() = EmailNotVerified;
 
-  factory AuthenticationState.success(UserEntity user) = _Success;
+  factory AuthenticationState.success(UserEntity user) = Success;
 
-  const factory AuthenticationState.failure() = _Failure;
+  const factory AuthenticationState.failure() = Failure;
 }