+++ lib/presentation/blocs/home/home_cubit.dart
@@ -1,3 +1,5 @@
+import 'package:assoshare/domain/entities/user/user_entity.dart';
+import 'package:assoshare/domain/repositories/user_repository.dart';
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
@@ -7,9 +9,12 @@ part 'home_state.dart';
 
 @injectable
 class HomeCubit extends Cubit<HomeState> {
-  HomeCubit() : super(const HomeState());
+  final UserRepository userRepository;
 
-  tabChanged(int index) {
-    emit(state.copyWith(tabIndex: index));
+  HomeCubit(this.userRepository) : super(const HomeState());
+
+  void initialize() {
+    final user = userRepository.getLocalUser();
+    emit(state.copyWith(user: user));
   }
 }