+++ lib/presentation/screens/ad_details_screen.dart
@@ -1,0 +1,13 @@
+import 'package:assoshare/domain/entities/ad/ad_entity.dart';
+import 'package:flutter/material.dart';
+
+class AdDetailsScreen extends StatelessWidget {
+  const AdDetailsScreen({super.key, required this.ad});
+
+  final AdEntity ad;
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}